move checks to /bin
