+++ bin/check-version.js
